+++ NEWS.md
@@ -1,5 +1,13 @@
 # News
 
+### 0.71.22
+
+* `tr_get_files()` is no longer specific to csv, tsv, and txt files.
+
+### 0.71.01
+
+* `tr_sigora_wrapper()` now runs quietly and returns NULL instead of an error
+
 ### 0.70.2
 
 * Added option to use `tr_sigora_wrapper()` with human or mouse input genes